vector_vertex_buffer を切り出し
