+++ font_rasterizer/src/lib.rs
@@ -16,3 +16,4 @@ mod screen_texture;
 mod screen_vertex_buffer;
 pub mod time;
 mod vector_vertex;
+mod vector_vertex_buffer;
